+++ packages/core/src/utils.ts
@@ -42,7 +42,8 @@ export function isPromise<T = any>(val: unknown): val is Promise<T> {
   return isObject(val) && isFunction(val.then) && isFunction(val.catch);
 }
 
-export const NO_OP = () => {};
+export const NO_OP = () => { };
+export const FALSE_OP = () => false;
 
 export function promisedTimeout(timeout: number): Promise<void> {
   return new Promise(res => {